webdriver: remove known web element cache
The known web element cache in the WebDriver test client, or
webdriver.Session._element_cache, is used only to avoid constructing
new webdriver.Element instances of the same web element and serves
no practical purpose beyond that

Since this client is intended for testing purposes, we would like
to be able to construct duplicate webdriver.Element instances,
so that e.g. fake elements can be constructed and send to the remote end.

Differential Revision: https://phabricator.services.mozilla.com/D8855

bugzilla-url: https://bugzilla.mozilla.org/show_bug.cgi?id=1499057
gecko-commit: 5df04b130a16afcaa917204055b8227f1ef0f411
gecko-integration-branch: autoland
gecko-reviewers: whimboo