[Background Fetch] Simplify fetch storage workflow.
In addition, MatchAll should return all records, including unprocessed
ones. To support this, the following changes were made:
- Creating a registration also stores all the requests with an empty
response in the cache.
- When an individual request is processed, the failure reason (if any)
is stored in the metadata.
- The logic of GetSettledFetchesTask was moved to
MarkRegistrationForDeletionTask. It checks the metadata rather than the
cache itself to find a failure reason (if any).
- Match/MatchAll logic was moved to a new database task
(MatchRequestsTask). A new API call was added to the cache storage to
allow querying request/response pairs. If a response is found to be
empty it will be exposed as a nullptr.

Change-Id: I631a3ef3da95117aed759a675fe591da5201eeca