[css-properties-values-api] Test styleMap in paint worklet.
Unfortunately, the existing test for this does too much behind a single
green square. If even a single assertion were to fail, it would hide the
result of all the other rests.

At the same time, adding 40something tests is perhaps a bit excessive, so
I've tried to reduce that number somewhat (mostly by only testing one
of the initial value cases, which I think is reasonable).

The aspect of the deleted test which verified that values were observable
from all API entry points (has, get, getAll, size and iterator) is now
a separate test as well.

R=ikilpatrick@chromium.org

Bug: 641877
Change-Id: Ie9316583ccef0ee4ddcaf688b7c1fac422649a39