Update Lato-Medium to 3.002.
This updates Lato-Medium from version 2.015 to version 3.002.
Lato-Medium is now available from https://github.com/latofonts and is
still available under the same OFL/SIL license.

In addition to many small fixes and tweaks this version adds a cmap
entry for the 'st' ligature with allows https://crrev.com/c/1294874 to
finally land. The update could have been made directly with that change,
but it is preferable to land this change directly.